+++ algorithms/algorithm_37bec7ux/src/ransac/occu.py
@@ -1,0 +1,5 @@
+# ground plane mask to occupancy grid
+
+# TODO: create a tool to tune grid paramters (scale, rotation, translation) in real time (or on a recording)
+
+
